changes in customer and sales order

+++ COMPROJ-3/src/main/java/com/punit/repository/ISalesOrderItemPersistence.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.punit.entities.SalesOrderItem;
 
 
-public interface ISalesOrderItemPersistence extends JpaRepository<SalesOrderItem, Long> {
+public interface ISalesOrderItemPersistence extends JpaRepository<SalesOrderItem, Integer> {
 			
 	
 }
